+++ sdk/core/core-client-rest/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 
+- Add logger in core-client-rest libraries.
+- Add a new property endpoint in ClientOptions and mark the baseUri as deprecated to encourage people to use endpoint.
+
 ### Breaking Changes
 
 ### Bugs Fixed